// src/modules/shared/types/sorting.ts
export enum OrderDirection {
  ASC = 'asc',
  DESC = 'desc'
}

export enum UserSortField {
  FULL_NAME = 'full_name',
  EMAIL = 'email',
  CREATED_AT = 'created_at'
}

export enum CategorySortField {
  NAME = 'name',
  CREATED_AT = 'created_at'
}

export enum ProductSortField {
  NAME = 'name',
  BARCODE = 'barcode',
  CREATED_AT = 'created_at'
}

export enum ProductEntrySortField {
  CREATED_AT = 'created_at',
  PRICE = 'price'
}

export enum StoreBrandSortField {
  NAME = 'name',
  CREATED_AT = 'created_at'
}

export enum StoreLocationSortField {
  ADDRESS = 'address',
  CREATED_AT = 'created_at',
  STORE_BRAND = 'store_brand'
}